added the main theme

+++ thinkmate/base/views.py
@@ -1,34 +1,123 @@
 from django.shortcuts import render,redirect
-from .models import Room,Topic
+from .models import Room,Topic,Message
+from django.db.models import Q
 from .forms import RoomForm
+from django.http import HttpResponse
+from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth.forms import UserCreationForm
+
+def LoginPage(request):
+    page='login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method=='POST':
+        username=request.POST.get('username').lower()
+        password=request.POST.get('password')
+
+        try:
+            user=User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist ")   
+        user=authenticate(request,username=username,password=password)    
+        if user:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,"password does not exist")
+
+    context={'page':page}
+    return render(request,'base/login_register.html',context)
+
+
+
+def LogoutPage(request):
+    logout(request)
+    return redirect('login')
+
+def RegisterPage(request):
+    form=UserCreationForm()
+    if request.method =='POST':
+        form=UserCreationForm(request.POST)
+        if form.is_valid():
+            user=form.save(commit=False)
+            user.username=user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,"An error occurred please try again")
+
+    return render(request,'base/login_register.html',{'form':form})
 
 
 
 
 def home(request):
-    rooms=Room.objects.all()
-    context={"rooms":rooms}
-    Topic=Topic.objects.all()
+    q=request.GET.get("q") if request.GET.get('q')!=None else ''
+    rooms=Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(description__icontains=q) 
+        )
+    room_messages=Message.objects.filter(Q(room__topic__name__icontains=q))
+    room_count=rooms.count()
+    topics=Topic.objects.all()
+    context={"rooms":rooms,"topics":topics,"room_count":room_count,"room_messages":room_messages}
     return render(request,"base/home.html",context)
+
+def userProfile(request,pk):
+    user=User.objects.get(id=pk)
+    rooms=user.room_set.all()
+    room_messages=user.message_set.all()
+    topics=Topic.objects.all()
+    context={"user":user,"rooms":rooms,"room_messages":room_messages,"topics":topics}
+    return render(request,"base/profile.html",context)
 
 def room(request,pk):
     room=Room.objects.get(id=pk)
-    
-    context={"room":room}
+    room_Messages=room.message_set.all()
+    participants=room.participants.all()
+    if request.method=="POST":
+        body=request.POST.get('body')
+        if body:
+            Message.objects.create(
+            user=request.user,
+            room=room,
+            body=body
+                    )
+            room.participants.add(request.user)
+            return redirect('room',pk=room.id)
+
+    context={"room":room,"room_Messages":room_Messages,"participants":participants}
     return render(request,"base/room.html",context)
+
+
+
+@login_required(login_url='login')
 def create_room(request):
     form=RoomForm()
     if request.method=="POST":
         form =RoomForm(request.POST)
         if form.is_valid():
-            form.save()
+            room=form.save(commit=False)
+            room.host=request.user
+            room.save()
             return redirect('home')
     context={'form':form}
     return render(request,"base/room_form.html",context)
 
+
+
+@login_required(login_url='login')
 def update_room(request,pk):
     room=Room.objects.get(id=pk)
     form=RoomForm(instance=room)
+    if request.user!=room.host:
+        return HttpResponse("You are not allowed to do this")
+
     if request.method=="POST":
         form=RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -36,10 +125,27 @@ def update_room(request,pk):
             return redirect("home")
     context={'form':form}
     return render(request,"base/room_form.html",context)
+
+@login_required(login_url='login')
 def delete_room(request,pk):
      
      room=Room.objects.get(id=pk)
+     if request.user!=room.host:
+        return HttpResponse("You are not allowed to do this")
      if request.method=="POST":
          room.delete()
          return redirect("home")
      return render(request,"base/delete.html",{'obj':room}) 
+
+
+
+@login_required(login_url="login")
+def deleteMessage(request,pk):
+    message=Message.objects.get(id=pk)
+
+    if request.user!=message.user:
+        return HttpResponse('you are not allowed here!!')
+    if request.method=="POST":
+        message.delete()
+        return redirect('home')
+    return render(request,'base/delete.html',{'obj':message})
